refactor(registry): Move manifest parsing logic from MCP to registry module
- Move ApiDocsUrl class, helper functions, and get_connector_api_docs_urls to airbyte/registry.py
- Simplify MCP get_api_docs_urls tool to call registry function
- Update _extract_docs_from_registry to use _get_registry_url() and include documentationUrl
- Add _DEFAULT_MANIFEST_URL constant to registry.py to avoid circular import
- Update tests to import from registry module
- All 8 unit tests passing, all lint checks passing
